Update package.json main to point to dist css file
package.json's `main` currently points to the non-existant dist/index.js file. The real main entry point for css-star-rating is the css file itself. By setting that to the main, it can be easily integrated with projects that support loading node modules (such as by using WebPack) as a simple `@import 'css-star-rating'`.